chore: adding proper exports for nodejs modules
