+++ package.json
@@ -3,13 +3,14 @@
   "version": "2.0.2-dev",
   "description": "awesome ascii charts",
   "main": "./dist/index.js",
-  "typings": "./dist/types/index.d.ts",
+  "types": "./dist/index.d.ts",
   "repository": {
     "type": "git",
     "url": "https://github.com/tool3/chartscii"
   },
   "scripts": {
     "build": "tsc",
+    "build:types": "npm-dts generate",
     "test": "mocha -r ts-node/register index.test.ts",
     "print": "node -e \"const keys = require('./snapshots/index.test'); Object.keys(keys).map(key => console.log(key + '\\n' + keys[key].join('\\n'))) \" "
   },
@@ -19,9 +20,6 @@
     "ascii",
     "chart"
   ],
-  "files": [
-    "/dist"
-  ],
   "author": "Tal Hayut",
   "license": "MIT",
   "devDependencies": {
@@ -29,6 +27,7 @@
     "@types/mocha": "10.0.6",
     "chai": "4.2.0",
     "mocha": "^10.0.0",
+    "npm-dts": "1.3.12",
     "shellfie": "^1.3.1",
     "snaptdout": "^2.1.0",
     "ts-node": "10.9.2",
